Split server.js and socket_handlers.js
